requirements.txt
GoogleAppEngineCloudStorageClient==1.8.3.1
django-allauth==0.16.1
django-appengine-toolkit==0.2.0
django-storages==1.1.8
djangotoolbox==1.6.2
oauthlib==0.6.1
python-openid==2.2.5
requests==2.2.1
requests-oauthlib==0.4.0
urllib3==1.8
wsgiref==0.1.2
